+++ src/applications/personalization/profile360/vet360/components/HomePhone.jsx
@@ -4,7 +4,7 @@ import PhoneField from './PhoneField';
 
 import {
   FIELD_NAMES
-} from '../../constants/vet360';
+} from '../constants';
 
 export default function HomePhone() {
   return (
