+++ src/applications/personalization/profile360/vet360/components/ResidentialAddress.jsx
@@ -4,7 +4,7 @@ import AddressField from './AddressField';
 
 import {
   FIELD_NAMES
-} from '../../constants/vet360';
+} from '../constants';
 
 export default function ResidentialAddress() {
   return (
