+++ src/applications/personalization/profile360/vet360/components/base/Transaction.jsx
@@ -5,7 +5,7 @@ import classNames from 'classnames';
 import {
   isPendingTransaction,
   isFailedTransaction
-} from '../../../util/transactions';
+} from '../../util/transactions';
 
 import Vet360TransactionInlineErrorMessage from './TransactionInlineErrorMessage';
 import Vet360TransactionPending from './TransactionPending';
