+++ src/applications/personalization/profile360/vet360/reducers/index.js
@@ -14,7 +14,7 @@ import {
 
 import {
   isFailedTransaction
-} from '../../util/transactions';
+} from '../util/transactions';
 
 const initialState = {
   modal: null,
