+++ src/applications/personalization/profile360/vet360/selectors.js
@@ -2,12 +2,12 @@ import backendServices from '../../../../platform/user/profile/constants/backend
 
 import {
   isVet360Configured
-} from '../util/local-vet360';
+} from './util/local-vet360';
 
 import {
   isFailedTransaction,
   isPendingTransaction
-} from '../util/transactions';
+} from './util/transactions';
 
 export function selectIsVet360AvailableForUser(state) {
   if (!isVet360Configured()) return true; // returns true if on localhost
